Add IDLEWorker.master-cmd-chan for communicating IDLE failures
If a user asks to idle on a nonexistant folder, the application will go
into a failure loop, likely resulting in a ban for the user. Instead of
filtering :idle/add|set commands by checking the server, it's easier and
more correct to simply destroy the IDLEWorker on FolderNotFoundException
failures.