Don't store the output of PassCmd (thanks dbasch!)
If the user gives us a plaintext password though, at least prevent it
from leaking into stacktraces by storing it as a byte array (thanks to
Justin Smith for the suggestion).

Unfortunately, this takes ~150ms per call to gpg on my machine. An
mbsyncrc with many PassCmds can be quite slow to parse (we cannot do it
in parallel either because of a bug in gpg-agent).