Add an optional password cache
This is a compromise.

On one hand, applications should store passwords in memory for as little
time as possible. On the other hand, the most likely program that will
likely handle our passwords, GnuPG + gpg-agent, has a bug that heavily
impacts performance:

    https://bugs.gnupg.org/gnupg/issue1190

Werner Koch:

> What happens here is that your IM software uses several connections to
> the gpg-agent and gpg-agent tries to answer them concurrently. This is
> in general a Good Thing.
>
> I can see that this is a bit annoying but it is a prety uncommon use
> case. In general you use different keys for different purposes. Or in
> your case better use OTR instad of OpenPGP. I can see how we can fix
> this but it will likely require a changed flow of control in gpg-agent
> which is an error prone change.
>
> Your IM software should use one connection for this task; running it
> in parallel has no advantage at all. Anyway, if it wants to optimize
> things it could watch out for the pinentry_launched inquiry and use
> this to detect whether it is likely that more pinentries will come up.
> But again, a lot of code for a questionable improvement. There is also
> a resource issue if it runs one conenction per avtive IM account.

Werner is, unfortunately, mistaken on one count: parallel calls to gpg
at one time drastically lowers the time taken to get multiple passwords
from multiple files.

A gpg decryption of a small bit of text takes on the order of 100-200ms
(depending on many factors). So for 5+ passwords there can be a full
second of cost to complete a task that may take half a second (like a
quick mbsync of a single folder with a single new message).

Parallelizing this reduces this cost to just over the time of a single
gpg call, therefore the penalty increases linearly with the number of
passwords.

I am personally unwilling to pay this penalty for a tiny bit of
security¹, so this cache is for the convenience of people like me.
There is an :app/clear Command that clears the cache, and a CLI option,
defaulting to NO cache.

¹ We cannot force the JVM to clear certain sensitive bits of memory,
  therefore there is always a relatively large amount of time during
  which passphrases will stay in active memory. This is vulnerable to
  attackers wielding gdb, swap files, cold boot attacks, and more.

  The extra security of leaving passphrases to the GC is merely that
  there is less opportunity for sensitive data leaking into swap files.
  All other attacks remain viable.

  WRT swap files, a user concerned about this should be using full disk
  encryption, not worrying about sensitive data in RAM.